pyproject.toml
[project]
name = "grok"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "dspy>=3.0.3",
    "pydantic>=2.11.9",
    "pyyaml>=6.0.2",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]

[tool.ruff]
line-length = 120
